parsing files added

+++ srcs/objects.c
@@ -6,7 +6,7 @@
 /*   By: pmartine <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/10/07 12:15:56 by pmartine          #+#    #+#             */
-/*   Updated: 2016/10/13 19:49:35 by pmartine         ###   ########.fr       */
+/*   Updated: 2016/10/15 14:34:50 by pmartine         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,12 +21,12 @@ void ft_init_sphere(t_obj *s)
 	s->color = RED;
 }
 
-int ft_sphere(t_obj *s, t_ray *r, float *d)
+int ft_sphere(t_obj *s, t_ray *r, double *d)
 {
-	float   b;
-	float   c;
-	float   discr;
-	t_vec  dist;
+	double	b;
+	double	c;
+	double	discr;
+	t_vec	dist;
 
 	dist = sub_vect(&r->pos, &s->pos);
 	b = scale(&r->dir, &dist);
@@ -51,11 +51,11 @@ void ft_init_plan(t_obj *s)
 	s->pos.z = 42;
 }
 
-int	ft_plan(t_obj *p, t_ray *r, float *d)
+int	ft_plan(t_obj *p, t_ray *r, double *d)
 {
-	float	num;
-	float	denum;
-	float	t;
+	double	num;
+	double	denum;
+	double	t;
 
 	num = p->d - scale(&p->dir, &r->pos);
 	denum = scale(&p->dir, &r->dir);
